requirements.txt
customtkinter
pystray
Pillow
psutil
keyboard
wmi
pywin32
